only require SYNC if we need to perform an inital git clone

+++ pym/_emerge/actions.py
@@ -1924,13 +1924,7 @@ def action_sync(settings, trees, mtimedb, myopts, myaction):
 	except OSError:
 		st = None
 
-	syncuri = settings.get("SYNC", "").strip()
-	if not syncuri:
-		writemsg_level("SYNC is undefined.\nPlease set SYNC to the remote location of the Portage repository.\n", noiselevel=-1, level=logging.ERROR)
-		return 1
-
 	os.umask(0o022)
-	dosyncuri = syncuri
 	updatecache_flg = False
 	if myaction == "metadata":
 		print("skipping sync")
@@ -1942,6 +1936,10 @@ def action_sync(settings, trees, mtimedb, myopts, myaction):
 		return 1
 	updatecache_flg = True
 	if not os.path.exists(myportdir+"/.git"):
+		syncuri = settings.get("SYNC", "").strip()
+		if not syncuri:
+			writemsg_level("SYNC is undefined.\nPlease set SYNC to the remote location of the Portage repository.\n", noiselevel=-1, level=logging.ERROR)
+			return 1
 		print(">>> Starting initial git clone with "+syncuri+"...")
 		try:
 			os.rmdir(myportdir)
